src/python/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
PyMuPDF==1.23.7
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
schedule==1.2.0
flask==2.3.3